pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
telebot = "^0.0.5"
aiogram = "^3.1.1"
keyboard = "^0.13.5"
requests = "^2.31.0"
scipy = "^1.11.4"
matplotlib = "^3.8.2"
pillow = "^10.1.0"
background = "^0.2.1"
flask = "^3.0.0"
numpy = "^1.26.2"
pytelegrambotapi = "^4.15.4"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"